requirements.txt
pandas
openpyxl
requests
python-dotenv
yfinance
google-generativeai
streamlit
pydantic
duckduckgo-search
